Add a way to lose the game
The game now keeps track of the number of mistakes you've made and the
letters that were wrong. This way the game knows what guesses you've
already made between your correct and incorrect guess, so you don't make
multiple of the same guess.

The game defaults to terminating at 6 mistakes, but this can be changed
in hangman.rb by changing the value set to the limit variable.